+++ packages/db/package.json
@@ -33,6 +33,6 @@
         "@onlook/eslint": "*",
         "drizzle-kit": "^0.31.4",
         "typescript": "^5.8.2",
-        "ai": "5.0.26"
+        "ai": "5.0.60"
     }
 }
